requirements.txt
djangorestframework
djangorestframework-simplejwt
drf_yasg
django_filter